Struct tor_rtcompat::async_std::AsyncStdRustlsRuntime

pub struct AsyncStdRustlsRuntime { /* private fields */ }

A Runtime powered by async_std and rustls.

Implementations

impl AsyncStdRustlsRuntime

pub fn create() -> IoResult<Self>

Return a new AsyncStdRustlsRuntime

Generally you should call this function only once, and then use Clone::clone() to create additional references to that runtime.

pub fn current() -> IoResult<Self>

Return an AsyncStdRustlsRuntime for the currently running async_std executor.

Note that since async_std executors are global, there is no distinction between this method and [AsyncStdNativeTlsRuntime::create()]: it is provided only for API consistency with the Tokio runtimes.