Struct tor_rtcompat::PreferredRuntime

pub struct PreferredRuntime { /* private fields */ }

The runtime that we prefer to use, out of all the runtimes compiled into the tor-rtcompat crate.

If tokio and async-std are both available, we prefer tokio for its performance. If native_tls and rustls are both available, we prefer native_tls since it has been used in Arti for longer.

Implementations

impl PreferredRuntime

pub fn current() -> Result<Self>

Obtain a PreferredRuntime from the currently running asynchronous runtime. Generally, this is what you want.

This tries to get a handle to a currently running asynchronous runtime, and wraps it; the returned PreferredRuntime isn’t the same thing as the asynchronous runtime object itself (e.g. tokio::runtime::Runtime).

Panics

When tor-rtcompat is compiled with the tokio feature enabled (regardless of whether the async-std feature is also enabled), panics if called outside of Tokio runtime context. See tokio::runtime::Handle::current.

Usage notes

Once you have a runtime returned by this function, you should just create more handles to it via Clone.

Limitations

If the tor-rtcompat crate was compiled with tokio support, this function will never return a runtime based on async_std.

pub fn create() -> Result<Self>

Create and return a new instance of the default Runtime.

Generally you should call this function at most once, and then use Clone::clone() to create additional references to that runtime.

Tokio users may want to avoid this function and instead obtain a runtime using PreferredRuntime::current: this function always builds a runtime, and if you already have a runtime, that isn’t what you want with Tokio.

If you need more fine-grained control over a runtime, you can create it using an appropriate builder type or function.