Trait tor_rtcompat::BlockOn

pub trait BlockOn: Clone + Send + Sync + 'static {
    // Required method
    fn block_on<F: Future>(&self, future: F) -> F::Output;
}

Trait for a runtime that can block on a future.

Required Methods

fn block_on<F: Future>(&self, future: F) -> F::Output

Run future until it is ready, and return its output.